### 2023-01-25 #5101 @danielskatz
@felixludos - thanks for your submission. As I'm not completely sure, the editors will now discussion if the submission is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS. This does not mean that we're discussing if it is software that is useful in research, just if JOSS considers it in scope for review as research software.  You should hear back in a week or two.

In the meantime, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
