### 2023-01-26 #5101 @felixludos
Thank you! If it helps, the clause "supports the functioning of research instruments or the execution of research experiments" seems the most relevant as I have been extensively using `omni-fig` for development and managing research experiments (in the field of machine learning), including managing jobs I submit to our in-house computing cluster.
