+++ paper.bib
@@ -116,6 +116,7 @@
   booktitle={International Conference on product-focused software process improvement},
   pages={153--167},
   year={2020},
+  doi={10.1007/978-3-030-64148-1_10},
   organization={Springer}
 }
 
@@ -142,6 +143,7 @@
   volume={212},
   pages={106622},
   year={2021},
+  doi={10.1016/j.knosys.2020.106622},
   publisher={Elsevier}
 }
 
