### 2023-02-08 #5101 @felixludos
Looking at the list, I think there are quite a few good matches. Pretty much anyone listing topics like: software engineering, infrastructure, configuration, automation, etc.
Such as ChristopherHaydenTodd, urjoshi, jarrah42, erik-whiting, idoby, julianpistorius, luciorq
