### 2023-04-07 #5350 @editorialbot
[REVIEW]: omni-fig: Unleashing Project Configuration and Organization in Python
**Submitting author:** @felixludos (<a href="http://orcid.org/0000-0002-3127-5707">Felix Leeb</a>)
**Repository:** https://github.com/felixludos/omni-fig
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.5
