### 2023-04-20 felixludos/omni-fig#2 @julianpistorius
Problem: Missing DOI
From https://github.com/openjournals/joss-reviews/issues/5350#issuecomment-1500681422

```
MISSING DOIs

- 10.1109/access.2022.3158675 may be a valid DOI for title: Reproducibility in Computing Research: An Empirical Study
```
