### 2023-05-02 #5350 @julianpistorius
@gkthiruvathukal I need guidance.

Under the Software Paper heading, the last task is:

> - [ ] **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

I noticed that @editorialbot found a missing DOI: https://github.com/openjournals/joss-reviews/issues/5350#issuecomment-1500681422

I created an issue for the author, but have had no response yet: https://github.com/felixludos/omni-fig/issues/2

Does this matter? Shall I consider this task complete even if there is a missing DOI?
