### 2023-05-03 felixludos/omni-fig#2 @julianpistorius
@felixludos this is apparently a prerequisite for me signing off on my review. See here: https://github.com/openjournals/joss-reviews/issues/5350#issuecomment-1532158241
