+++ paper.bib
@@ -74,6 +74,7 @@
   volume={10},
   pages={29207--29223},
   year={2022},
+  doi={10.1109/access.2022.3158675},
   publisher={IEEE}
 }
 
