### 2023-05-03 felixludos/omni-fig#2 @felixludos
The missing DOI should be fixed now.
Sorry for the delay!
