+++ tests/test_config.py
@@ -605,13 +605,16 @@ def test_pull_cycle():
 
 def test_underscores():
 
+	# NOTE: dashes default to underscores (but not vice versa)
+	# so, in general, queries should use dashes, configs should use underscores
+
 	A = fig.create_config(a_b=1, _b=2, **{'-c': 3, 'd-e': 4, '--f': 5})
 
 	assert A.pull('a_b') == 1
 	assert A.pull('_b') == 2
-	assert A.pull('_c') == 3
-	assert A.pull('d_e') == 4
-	assert A.pull('__f') == 5
+	# assert A.pull('_c') == 3
+	# assert A.pull('d_e') == 4
+	# assert A.pull('__f') == 5
 
 	assert A.pull('a-b') == 1
 	assert A.pull('-b') == 2
@@ -621,6 +624,28 @@ def test_underscores():
 
 	print()
 	print(A)
+
+
+def test_underscores2():
+	data = {'-x':1, '_y': '-a_'}
+	A = fig.create_config(a_b=data, _b=2, **{'-c': data, 'd-e': 4, '--f': 5})
+
+	assert A.pull('a_b') == data
+	assert A.pull('a-b') == data
+	assert A.pull('-c') == data
+
+	assert A.pull('a_b.-x') == 1
+	assert A.pull('a_b._y') == '-a_'
+
+	assert A.pull('a-b.-x') == 1
+	assert A.pull('a-b._y') == '-a_'
+
+	assert A.pull('-c.-x') == 1
+	assert A.pull('-c.-y') == '-a_'
+
+
+
+
 
 
 
