### 2023-09-10 3142eff
removed public my_config
