### 2023-09-11 3729b69
convenience printing function
