+++ requirements.txt
@@ -1,4 +1,4 @@
 pyyaml
 toml
-omnibelt>=0.7.3
+omnibelt>=0.8.0
 tabulate
