+++ .readthedocs.yml
@@ -7,6 +7,8 @@ version: 2
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
+  builder: html
+  fail_on_warning: false
   configuration: docs/conf.py
 
 # Build documentation with MkDocs
