+++ .readthedocs.yml
@@ -20,16 +20,16 @@ formats:
   - pdf
 
 # Optionally set the version of Python and requirements required to build your docs
-#python:
-#  version: 3.8
-#  install:
-#    - requirements: docs/requirements.txt
+python:
+  version: 3.8
+  install:
+    - requirements: docs/requirements.txt
 
 
 build:
   os: "ubuntu-22.04"
   tools:
     python: "3.8"
-  commands:
-    - pip install -r docs/requirements.txt
+#  commands:
+#    - pip install -r docs/requirements.txt
 #    - pip install .
