+++ .readthedocs.yml
@@ -1,35 +1,72 @@
-# .readthedocs.yml
-# Read the Docs configuration file
+## .readthedocs.yml
+## Read the Docs configuration file
+## See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
+#
+## Required
+#version: 2
+#
+## Build documentation in the docs/ directory with Sphinx
+#sphinx:
+#  builder: html
+#  fail_on_warning: false
+#  configuration: docs/conf.py
+#
+## Build documentation with MkDocs
+##mkdocs:
+##  configuration: mkdocs.yml
+#
+## Optionally build your docs in additional formats such as PDF
+#formats:
+#  - pdf
+#
+## Optionally set the version of Python and requirements required to build your docs
+#python:
+#  version: 3.8
+#  install:
+#    - requirements: docs/requirements.txt
+#
+#
+#build:
+#  os: "ubuntu-22.04"
+#  tools:
+#    python: "3.8"
+##  commands:
+##    - pip install -r docs/requirements.txt
+##    - pip install .
+
+# Read the Docs configuration file for Sphinx projects
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
 # Required
 version: 2
 
-# Build documentation in the docs/ directory with Sphinx
+# Set the OS, Python version and other tools you might need
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+    # You can also specify other tool versions:
+    # nodejs: "20"
+    # rust: "1.70"
+    # golang: "1.20"
+
+# Build documentation in the "docs/" directory with Sphinx
 sphinx:
-  builder: html
-  fail_on_warning: false
   configuration: docs/conf.py
+  # You can configure Sphinx to use a different builder, for instance use the dirhtml builder for simpler URLs
+  # builder: "dirhtml"
+  # Fail on all warnings to avoid broken references
+  # fail_on_warning: true
 
-# Build documentation with MkDocs
-#mkdocs:
-#  configuration: mkdocs.yml
-
-# Optionally build your docs in additional formats such as PDF
+# Optionally build your docs in additional formats such as PDF and ePub
 formats:
   - pdf
+#    - epub
 
-# Optionally set the version of Python and requirements required to build your docs
+# Optional but recommended, declare the Python requirements required
+# to build your documentation
+# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
-  version: 3.8
   install:
     - requirements: docs/requirements.txt
 
-
-build:
-  os: "ubuntu-22.04"
-  tools:
-    python: "3.8"
-#  commands:
-#    - pip install -r docs/requirements.txt
-#    - pip install .
