+++ .readthedocs.yml
@@ -1,39 +1,3 @@
-## .readthedocs.yml
-## Read the Docs configuration file
-## See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-#
-## Required
-#version: 2
-#
-## Build documentation in the docs/ directory with Sphinx
-#sphinx:
-#  builder: html
-#  fail_on_warning: false
-#  configuration: docs/conf.py
-#
-## Build documentation with MkDocs
-##mkdocs:
-##  configuration: mkdocs.yml
-#
-## Optionally build your docs in additional formats such as PDF
-#formats:
-#  - pdf
-#
-## Optionally set the version of Python and requirements required to build your docs
-#python:
-#  version: 3.8
-#  install:
-#    - requirements: docs/requirements.txt
-#
-#
-#build:
-#  os: "ubuntu-22.04"
-#  tools:
-#    python: "3.8"
-##  commands:
-##    - pip install -r docs/requirements.txt
-##    - pip install .
-
 # Read the Docs configuration file for Sphinx projects
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
@@ -44,7 +8,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.11"
+    python: "3.8"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
