+++ omnifig/__init__.py
@@ -5,7 +5,7 @@ from .registration import script, component, modifier, creator, autocomponent, a
 from .top import get_current_project, get_project, switch_project, iterate_projects, \
 	entry, main, run_script, run, quick_run, initialize, create_config, parse_argv
 
-from .config import ConfigNode as Node
+from .config import ConfigNode as Configuration
 from . import exporting as _exporting
 from .behaviors import Behavior
 
