+++ tests/test_running.py
@@ -12,7 +12,7 @@ def test_initialize_empty():
 
 	default = fig.get_current_project()
 
-	assert default.name == 'omni-fig'
+	assert default.name is None
 	assert not default.is_activated
 
 	assert len(list(profile.iterate_projects())) == 1
@@ -31,7 +31,7 @@ def test_initialize():
 
 	default = fig.get_current_project()
 
-	assert default.name == 'omni-fig'
+	assert default.name is None
 	assert not default.is_activated
 
 	fig.initialize('example2')
@@ -44,7 +44,7 @@ def test_initialize():
 	assert default.is_activated
 
 	assert len(list(default.iterate_components())) == 0
-	assert fig.get_current_project().name == 'omni-fig'
+	assert fig.get_current_project().name is None
 
 	assert proj.name == 'example2'
 	assert len(list(proj.iterate_components())) == 4
@@ -56,7 +56,7 @@ def test_initialize_multi():
 
 	default = fig.get_current_project()
 
-	assert default.name == 'omni-fig'
+	assert default.name is None
 	assert not default.is_activated
 
 	assert len(list(profile.iterate_projects())) == 1
