+++ ignore.omnifig.yaml
@@ -1,0 +1,2 @@
+
+py_info: 'omnifig/_info.py'
