+++ .omnifig.yaml
@@ -1,2 +1,0 @@
-
-py_info: 'omnifig/_info.py'
