### 2024-02-15 7626b68
fixed special dependencies
