### 2024-03-25 #5350 @jarrah42
There are two minor problems.

1. The community guidelines are very brief. I would be good to provide more information on how to contribute to the project, how to obtain support, etc.
2. There is a docs:failing badge on the GitHub site.

Other than these I think it is fine.
