### 2024-03-26 16b6305
Update README with contribution information and fixed paper typos
