+++ paper.md
@@ -24,33 +24,33 @@ bibliography: paper.bib
 
 We present a lightweight package to take care of the configuration and organization of all your Python projects.
 Although `omni-fig` is well suited for projects of virtually any size or complexity, it is specifically designed for research projects where a small development team doesn't want to waste time on boilerplate code or a bespoke user interface.
-Nevertheless, the intuitive project structure encourages developers good design practices making it easier for collaborators and prospective users to rapidly find and understand the core contributions they need.
+Nevertheless, the intuitive project structure encourages developers to good design practices making it easier for collaborators and prospective users to rapidly find and understand the core contributions they need.
 The feature-rich configuration system completely abstracts the implementation from the specification of parameters, so the developer can focus entirely on creating new functionality, while the user can quickly and precisely specify the desired functionality by composing modular config files, through the command line interface, or even in an interactive environment like Jupyter.
-`omni-fig` transforms the project organization and configuration from a bothersome distraction into a joy, thereby both improving clarity while also accelerating development.
+`omni-fig` transforms the project organization and configuration from a bothersome distraction into a joy, thereby improving clarity while also accelerating development.
 
 # Introduction
 
 One particularly promising trend in computational research, especially in machine learning, is that releasing the code associated with projects is becoming increasingly common [@pineau2021improving]. Not only is this crucial for reproducibility [@guyonartificial], but this also fosters interdisciplinary research and progress in industry [@aho2020demystifying].
 
 However, the needs of developers differ significantly from those of potential users. *Developers* (and especially researchers) prefer flexible, open-ended tools tailored for extensibility to prototype and synthesize novel methods. Meanwhile, *users* prefer tools that automate details and provide a simple interface that is digestible and directly applicable to some downstream problem. Even fellow researchers exploring new methods want to quickly understand the essence of the contribution rather than getting bogged down in some idiosyncrasies of a project.
-For example, in machine learning, this has given rise to a variety of development practices [@ebert2016devops; @treveil2020introducing] and AutoML packages [@he2021automl] to make cutting edge methods more highly accessible for a variety of real-world applications. However, the opaque rigid interfaces and product-oriented focus caters more to the end-user, thereby increasing friction in design and synthesis for developers.
+For example, in machine learning, this has given rise to a variety of development practices [@ebert2016devops; @treveil2020introducing] and AutoML packages [@he2021automl] to make cutting edge methods more highly accessible for a variety of real-world applications. However, the opaque rigid interfaces and product-oriented focus cater more to the end-user, thereby increasing friction in design and synthesis for developers.
 
 # Statement of Need
 
-Project organization and configuration is an important albeit unglamourous part of every project. Neglecting this aspect can significantly impede prototyping and development as well as making the project less understandable and useable for any potential user or collaborator. Consequently, there exist quite a few packages already to help with this task.
+Project organization and configuration is an important albeit unglamorous part of every project. Neglecting this aspect can significantly impede prototyping and development as well as making the project less understandable and useable for any potential user or collaborator. Consequently, there exist quite a few packages already to help with this task.
 
[…]
