### 2024-04-01 1560b4d
enable function components
