### 2024-05-04 #5350 @gkthiruvathukal
@felixludos Thank you for your patience.  Please do the following:

- Make a tagged release of your software, and list the version tag of the archived version here as a separate reply.
- Archive the reviewed software in Zenodo
- Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
- List the Zenodo DOI of the archived version here as a separate reply.
