### 2024-06-02 #5350 @felixludos
The release tag is v1.0.5
