### 2024-06-02 #5350 @felixludos
I also uploaded the release to Zenodo with DOI [10.5281/zenodo.11424101](https://zenodo.org/doi/10.5281/zenodo.11424101)

Sorry it took so long.
