### 2024-06-04 #5350 @danielskatz
@felixludos - please confirm that this proof looks correct to you.
