### 2024-06-04 #5350 @felixludos
@danielskatz The proof looks correct to me! Thanks!
@gkthiruvathukal Regarding the missing DOIs, some references don't seem to have any DOI (e.g. other github repos that haven't been archived). Is it ok to leave the references as is (preferable) or is there another workaround? Thanks!
