### 2024-06-04 #5350 @danielskatz
The references without DOIs are fine.  And the one suggestion from editorialbot is wrong.

I'll proceed to accept at this point.
