### 2024-06-04 #5350 @danielskatz
Sorry, I missed the fact that one of the reviewers should have been deleted, and shouldn't appear in the final paper - I changed this after the accept, so now need to re-accept to update the deposited paper.

@gkthiruvathukal - please be sure to remove any reviewers that stop working on a paper via the `remove . as reviewer` command next time.
