### 2024-06-04 #5350 @danielskatz
Congratulations to @felixludos (Felix Leeb) on your publication!!

JOSS depends on volunteers and couldn't be successful without your work
